Make the docs agree with the hook the code actually uses
Ninth review pass.

Three comments still said the OSC 2 title is cleared at command start, including
the KDoc on windowTitleFlow that a consumer reads to learn what empty means. The
code clears at prompt start and argues at length for it, so a reader found both
claims and had to work out which was true. Fixed, and the root cause with it:
that argument existed in several copies, so AGENTS.md is now the one that
carries it and the call site keeps the decision plus a pointer.

Also from the review:
- resolveWindowTitle and notificationTitle move out of the 2.8k-line composable
  file into WindowTitle.kt. They are title policy, not part of the composable.
- the local NotificationTitleProvider val no longer shadows the imported
  notificationTitle function, and the AtomicReference import is no longer
  spliced into the middle of the ai.rever block.
- NotificationTitleProvider is internal, with a test pinning what it answers
  before its tab is attached: the app's OSC 2 title if there is one, otherwise
  the fallback, which is exactly pre-PR behaviour.
- runCatching becomes try/catch (e: Exception), so a CancellationException
  crossing the LaunchedEffect cannot be swallowed and reported as a style
  failure.
- the TerminalSettings KDoc no longer links an overloaded name Dokka would warn
  on as ambiguous.

Two findings are hand-checks in the running app rather than code, and are on the
list for the manual pass: whether the title bar macOS reveals on a top-edge
hover in fullscreen swallows clicks on the first tab row, and whether a TUI that
rewrites its title continuously makes the window title churn.
