+++ compose-ui/src/desktopMain/kotlin/ai/rever/bossterm/compose/settings/TerminalSettings.kt
@@ -171,7 +171,7 @@ data class TerminalSettings(
      * Changing this requires app restart to take effect.
      *
      * "No transparency" is an AWT restriction, not a macOS one - see
-     * [ai.rever.bossterm.compose.window.applyFullWindowContent], which records what was measured.
+     * `applyFullWindowContent` in `window/NativeTitleBarStyle.kt`, which records what was measured.
      * Short version: AWT gives an alpha-capable backing store only to windows it treats as
      * translucent and refuses that for decorated frames, and forcing the NSWindow non-opaque
      * underneath does not help because the surface has no alpha channel. Do not spend an
